# doc/OpenAPI.json
{
  "openapi": "3.1.0",
...
  "info": {
    "title": "K-XYZ 智能餐盒全链路 API",
    "version": "4.4.0",
    "description": "根据 D:\\MIT\\newdev\\doc\\API.md（v4.4版）整理生成的 OpenAPI 描述。除健康检查、硬件遥测上报、注册、登录外，其余接口默认需要 Bearer JWT。"
  },
  "servers": [
    {
      "url": "https://api.mit.chenyuxia.com/api/v1",
      "description": "生产环境"
    },
    {
      "url": "http://127.0.0.1:8080/api/v1",
      "description": "本地开发"
    }
  ],
...
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/meals/{meal_id}": {
      "get": {
        "tags": [
          "Meals"
        ],
        "summary": "获取单次用餐高精详情",
        "operationId": "getMealDetail",
        "parameters": [
          {
            "name": "meal_id",
            "in": "path",
            "required": true,
            "description": "用餐 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "用餐详情查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealDetailResponse"
                },
                "example": {
                  "meal_id": "str",
                  "grid_details": [
                    {
                      "grid_index": 1,
                      "food_name": "糙米饭",
                      "served_g": 200,
                      "intake_g": 150,
                      "total_cal": 174,
                      "speed_g_per_min": 6.0
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/meals/{meal_id}/trajectory": {
      "get": {
        "tags": [
          "Meals"
        ],
        "summary": "获取就餐时序轨迹",
        "description": "支持增量查询与降采样。",
        "operationId": "getMealTrajectory",
        "parameters": [
          {
            "name": "meal_id",
            "in": "path",
            "required": true,
            "description": "用餐 ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "last_timestamp",
            "in": "query",
            "required": false,
            "description": "增量查询基准时间戳。传入上一次返回结果中的 last_timestamp 后，接口仅返回该时间点之后的新增轨迹数据。",
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "sample_interval",
            "in": "query",
            "required": false,
            "description": "降采样时间间隔，单位为秒。",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "轨迹查询成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MealTrajectoryResponse"
                },
                "example": {
                  "meal_id": "string",
                  "items": [
                    {
                      "timestamp": "2026-03-14T09:00:01Z",
                      "weights": {
                        "grid_1": 195.0,
                        "grid_2": 80.0,
                        "grid_3": 120.0,
                        "grid_4": 50.0
                      }
                    }
                  ],
                  "last_timestamp": "2026-03-14T09:00:01Z"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "除 /ping、/hardware/telemetry、/auth/register、/auth/login 外，其余接口默认需要 Bearer Token。"
      }
    },
    "schemas": {
...
      "GridWeights": {
        "type": "object",
        "properties": {
          "grid_1": {
            "type": "number"
          },
          "grid_2": {
            "type": "number"
          },
          "grid_3": {
            "type": "number"
          },
          "grid_4": {
            "type": "number"
          }
        },
        "required": [
          "grid_1",
          "grid_2",
          "grid_3",
          "grid_4"
        ]
      },
...
      "MealDetailResponse": {
        "type": "object",
        "properties": {
          "meal_id": {
            "type": "string"
          },
          "grid_details": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealGridDetail"
            }
          }
        },
        "required": [
          "meal_id",
          "grid_details"
        ]
      },
      "MealGridDetail": {
        "type": "object",
        "properties": {
          "grid_index": {
            "type": "integer"
          },
          "food_name": {
            "type": "string"
          },
          "served_g": {
            "type": "number"
          },
          "intake_g": {
            "type": "number"
          },
          "total_cal": {
            "type": "number"
          },
          "speed_g_per_min": {
            "type": "number"
          }
        },
        "required": [
          "grid_index",
          "food_name",
          "served_g",
          "intake_g",
          "total_cal",
          "speed_g_per_min"
        ]
      },
      "MealTrajectoryResponse": {
        "type": "object",
        "properties": {
          "meal_id": {
            "type": "string"
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/MealTrajectoryItem"
            }
          },
          "last_timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "meal_id",
          "items",
          "last_timestamp"
        ]
      },
      "MealTrajectoryItem": {
        "type": "object",
        "properties": {
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "weights": {
            "$ref": "#/components/schemas/GridWeights"
          }
        },
        "required": [
          "timestamp",
          "weights"
        ]
      },
...
    }
  }
}